requirements.txt
asgiref==3.4.1
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
cryptography==3.4.7
Django==3.2.6
djangorestframework==3.12.4
idna==3.2
pycparser==2.20
pyOpenSSL==20.0.1
pytz==2021.1
redis==3.5.3
requests==2.26.0
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.6